Removed duplicate db files
